backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
curl-cffi
langchain
langchain-openai
langchain-core
langchain-community
langgraph
celery[sqlalchemy]
redis
tavily-python
pythonnet
python-multipart
hubspot-api-client
pandas
python-dotenv
alembic
pydantic[email]
aiosmtplib
ddgs
requests
pytz